(svn r22131) -Fix: Spectators had trouble closing buoy windows. (Terkhen)

+++ src/waypoint_gui.cpp
@@ -79,8 +79,12 @@ public:
 	~WaypointWindow()
 	{
 		Owner owner = this->owner;
+
+		/* Buoys have no owner and can be used by everyone. Show only 'our' vehicles */
 		if (!Company::IsValidID(owner)) owner = _local_company;
-		DeleteWindowById(GetWindowClassForVehicleType(this->vt), VehicleListIdentifier(VL_STATION_LIST, this->vt, owner, this->window_number).Pack(), false);
+
+		/* Well, spectators otoh */
+		if (Company::IsValidID(owner)) DeleteWindowById(GetWindowClassForVehicleType(this->vt), VehicleListIdentifier(VL_STATION_LIST, this->vt, owner, this->window_number).Pack(), false);
 	}
 
 	virtual void SetStringParameters(int widget) const
